refactor: MenuService 내부 Product import 제거
