+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,16 +1,12 @@
 package kitchenpos.menu.application;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.menu.dao.MenuDao;
-import kitchenpos.menu.dao.MenuGroupDao;
 import kitchenpos.menu.dao.MenuProductDao;
-import kitchenpos.product.dao.ProductDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.common.domain.Price;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.MenuValidator;
 import kitchenpos.menu.dto.request.MenuRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
 import org.springframework.stereotype.Service;
@@ -20,18 +16,15 @@ import org.springframework.transaction.annotation.Transactional;
 public class MenuService {
 
     private final MenuDao menuDao;
-    private final MenuGroupDao menuGroupDao;
     private final MenuProductDao menuProductDao;
-    private final ProductDao productDao;
+    private final MenuValidator menuValidator;
 
     public MenuService(final MenuDao menuDao,
-                       final MenuGroupDao menuGroupDao,
                        final MenuProductDao menuProductDao,
-                       final ProductDao productDao) {
+                       final MenuValidator menuValidator) {
         this.menuDao = menuDao;
-        this.menuGroupDao = menuGroupDao;
+        this.menuValidator = menuValidator;
         this.menuProductDao = menuProductDao;
-        this.productDao = productDao;
     }
 
     @Transactional
@@ -39,8 +32,7 @@ public class MenuService {
         final Menu menu = menuRequest.toMenu();
         final List<MenuProduct> menuProducts = menuRequest.toMenuProducts();
 
-        validateMenuGroup(menu);
-        validateMenuPrice(menu.getPrice(), calculateSumOfMenuPrice(menuProducts));
+        menuValidator.validate(menu, menuProducts);
 
         final Menu savedMenu = menuDao.save(menu);
         final List<MenuProduct> savedMenuProducts = saveMenuProducts(menuProducts, savedMenu);
@@ -53,32 +45,6 @@ public class MenuService {
                 .map(menuProduct -> menuProductDao.save(
                         new MenuProduct(savedMenu.getId(), menuProduct.getProductId(), menuProduct.getQuantity())))
                 .collect(Collectors.toList());
-    }
-
-    public void validateMenuPrice(Price menuPrice, BigDecimal sum) {
-        if (menuPrice.getValue().compareTo(sum) > 0) {
-            throw new IllegalArgumentException("메뉴의 가격은 상품 수량 * 상품 가격의 합보다 클 수 없습니다.");
-        }
-    }
-
-    private BigDecimal calculateSumOfMenuPrice(final List<MenuProduct> menuProducts) {
-        BigDecimal sum = BigDecimal.ZERO;
-        for (final MenuProduct menuProduct : menuProducts) {
-            final Product product = getProduct(menuProduct);
-            sum = sum.add(product.multiply(menuProduct.getQuantity()));
-        }
-        return sum;
-    }
-
-    private Product getProduct(final MenuProduct menuProduct) {
-        return productDao.findById(menuProduct.getProductId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 상품입니다."));
-    }
-
-    private void validateMenuGroup(final Menu menu) {
-        if (!menuGroupDao.existsById(menu.getMenuGroupId())) {
-            throw new IllegalArgumentException("존재하지 않는 메뉴 그룹입니다.");
-        }
     }
 
     public List<MenuResponse> list() {
